Consolidate `wait_closed` into `BaseSerialTransport`
